api/requirements.txt
tensorflow==2.15.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.2.0
requests==2.31.0
numpy==1.26.4
matplotlib==3.8.3